Created Medicine -> Model, Controller, Service, Repository, Command,Handler and tests

+++ src/building blocks/PetGuadian.Application/Dto/PetDto/CreatePetDto.cs
@@ -12,6 +12,6 @@ namespace PetGuadian.Application.Dto.PetDto
         char Gender, 
         AnimalSpecies Specie,
         DateTime BirthDate,
-        long? Weight,
+        float? Weight,
         Guid UserId);
 }
